requirements.txt
numpy==1.26.4
paddlepaddle==3.0.0
paddleocr==2.9.1
opencv-python==4.9.0.80
Pillow==12.1.1
scikit-image==0.22.0
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.20